configure environments in webpack; convert webpack config to typescript
