+++ @types/custom.d.ts
@@ -1,6 +1,20 @@
 /* eslint-disable import/no-default-export -- enable default exports */
 
 /**
+ * Type definitions for import.meta.env
+ */
+
+/** can be extended in package to contain specific keys */
+interface ImportMetaEnv {
+  readonly APP_REACT_TITLE: string;
+  [string]: string;
+}
+
+interface ImportMeta {
+  readonly env: ImportMetaEnv;
+}
+
+/**
  * # Custom
  *
  * See: https://webpack.js.org/guides/typescript/#importing-other-assets
