+++ apps/app-vite/package.json
@@ -5,7 +5,9 @@
   "type": "module",
   "scripts": {
     "dev": "vite",
+    "__dev:test": "NODE_ENV=development vite",
     "build": "tsc && vite build",
+    "__build:test": "NODE_ENV=development tsc && NODE_ENV=development vite build",
     "build-verbose": "tsc && DEBUG=vite-tsconfig-paths vite build",
     "lint": "eslint . --ext ts,tsx --report-unused-disable-directives --max-warnings 0",
     "preview": "vite preview",
