+++ apps/app-webpack/package.json
@@ -4,8 +4,8 @@
   "private": true,
   "type": "module",
   "scripts": {
-    "dev": "webpack serve --config config/webpack.config.js",
-    "build": "webpack --config config/webpack.config.js",
+    "dev": "NODE_OPTIONS='--loader ts-node/esm' NODE_ENV=development webpack serve --mode development --config config/webpack.config.ts",
+    "build": "NODE_OPTIONS='--loader ts-node/esm' NODE_ENV=production webpack --mode production --config config/webpack.config.ts",
     "test": "echo 'Add test script here'",
     "lint": "eslint .",
     "compile": "tsc",
@@ -19,14 +19,18 @@
   "devDependencies": {
     "@repo/eslint-config": "workspace:*",
     "@repo/typescript-config": "workspace:*",
+    "@types/dotenv-webpack": "^7.0.7",
     "@types/node": "^22.7.4",
     "@types/react": "^18.3.11",
     "@types/react-dom": "^18.3.0",
+    "@types/webpack-dev-server": "^4.7.2",
     "copy-webpack-plugin": "^12.0.2",
     "css-loader": "^7.1.2",
+    "dotenv-webpack": "^8.1.0",
     "html-webpack-plugin": "^5.6.0",
     "mini-css-extract-plugin": "^2.9.1",
     "ts-loader": "^9.5.1",
+    "ts-node": "^10.9.2",
     "tsconfig-paths-webpack-plugin": "^4.1.0",
     "webpack": "^5.95.0",
     "webpack-cli": "^5.1.4",
