+++ packages/app-core/src/index.tsx
@@ -17,12 +17,11 @@ export function AppCore(props: {
 }): React.JSX.Element {
 
   // console.log("* process.env:", process.env);
-  // console.log("* process.env.API_URL:", process.env.API_URL);
+  // console.log("* process.env.APP_REACT_TITLE:", process.env.APP_REACT_TITLE);
+  // console.log("* process.env.HOME:", process.env.HOME);
 
-  console.log("* import.meta.env:", import.meta.env);
-
-  // const t = import.meta.env.APP_REACT_TITLE;
-
+  // console.log("* import.meta.env:", import.meta.env);
+  // console.log("* import.meta.env.HOME:", import.meta.env.HOME);
   console.log("* import.meta.env.APP_REACT_TITLE:", import.meta.env.APP_REACT_TITLE);
 
   return (
